refactor: hide executor proxies
