+++ src/promptflow-devkit/promptflow/_proxy/__init__.py
@@ -2,7 +2,8 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
+from ._base_executor_proxy import AbstractExecutorProxy
 from ._base_inspector_proxy import AbstractInspectorProxy
 from ._proxy_factory import ProxyFactory
 
-__all__ = ["ProxyFactory", "AbstractInspectorProxy"]
+__all__ = ["ProxyFactory", "AbstractInspectorProxy", "AbstractExecutorProxy"]
