+++ src/promptflow-devkit/promptflow/_proxy/_csharp_executor_proxy.py
@@ -12,8 +12,9 @@ from typing import NoReturn, Optional
 from promptflow._core._errors import UnexpectedError
 from promptflow._sdk._constants import OSType
 from promptflow._utils.flow_utils import is_flex_flow
-from promptflow.batch._csharp_base_executor_proxy import CSharpBaseExecutorProxy
 from promptflow.storage._run_storage import AbstractRunStorage
+
+from ._csharp_base_executor_proxy import CSharpBaseExecutorProxy
 
 EXECUTOR_SERVICE_DOMAIN = "http://localhost:"
 EXECUTOR_SERVICE_DLL = "Promptflow.dll"
