+++ src/promptflow-devkit/promptflow/batch/__init__.py
@@ -3,19 +3,10 @@
 # ---------------------------------------------------------
 
 # flake8: noqa
-from ._base_executor_proxy import AbstractExecutorProxy, APIBasedExecutorProxy
 from ._batch_engine import BatchEngine
-from ._csharp_base_executor_proxy import CSharpBaseExecutorProxy
-from ._csharp_executor_proxy import CSharpExecutorProxy
-from ._python_executor_proxy import PythonExecutorProxy
 from ._result import BatchResult
 
 __all__ = [
-    "AbstractExecutorProxy",
-    "APIBasedExecutorProxy",
     "BatchEngine",
-    "CSharpExecutorProxy",
-    "CSharpBaseExecutorProxy",
-    "PythonExecutorProxy",
     "BatchResult",
 ]
