+++ src/promptflow-devkit/promptflow/batch/_batch_engine.py
@@ -19,7 +19,8 @@ from promptflow._constants import (
     MessageFormatType,
 )
 from promptflow._core._errors import ResumeCopyError, UnexpectedError
-from promptflow._proxy import ProxyFactory
+from promptflow._proxy import AbstractExecutorProxy, ProxyFactory
+from promptflow._proxy._python_executor_proxy import PythonExecutorProxy
 from promptflow._utils.async_utils import async_run_allowing_running_loop
 from promptflow._utils.context_utils import _change_working_dir
 from promptflow._utils.execution_utils import (
@@ -41,10 +42,8 @@ from promptflow._utils.utils import (
     transpose,
 )
 from promptflow._utils.yaml_utils import load_yaml
-from promptflow.batch import AbstractExecutorProxy
 from promptflow.batch._batch_inputs_processor import BatchInputsProcessor
 from promptflow.batch._errors import BatchRunTimeoutError
-from promptflow.batch._python_executor_proxy import PythonExecutorProxy
 from promptflow.batch._result import BatchResult
 from promptflow.contracts.flow import Flow
 from promptflow.contracts.run_info import FlowRunInfo, Status
